Fix g_ourPitch to 16bit value as it's load bearing

+++ src/egdata.c
@@ -1838,7 +1838,7 @@ struct BulletTrack bulletTracks[20];
 /* 3x3 rotation matrix scratch buffers (9 words): written by
  * multiplyMatrix3x3Far/buildRotationMatrixFar, read as [axis]/[3+axis]/[6+axis]. */
 int16 g_matrixScratch[9];
-int g_ourPitch = 0;
+int16 g_ourPitch = 0;
 /* 16-bit word-degree angles: arithmetic must wrap at 16 bits (e.g. 0x10000 - roll,
    -head). Keep these int16 so assignments wrap on store as the DOS build did. */
 int16 g_ourRoll = 0;
